backend/requirements.txt
fastapi==0.116.1
grad_cam==1.5.5
Pillow==11.3.0
pydantic==1.10.22
torch==2.7.1
torchvision==0.22.1
uvicorn[standard]>=0.30
python-multipart
numpy
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub